# src/main/resources/static/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Rest API
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/v1/cards/{id}/confirm-block:
    put:
      tags:
        - card-controller
      operationId: confirmBlockRequest
      summary:
        "Confirm card block (ADMIN)"
      description: "Block card if the status is REQUESTED_BLOCK"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
  /api/v1/cards/{id}/block:
    put:
      tags:
        - card-controller
      operationId: blockCard
      summary:
        "Block card (ADMIN)"
      description: "Change card status to BLOCKED"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
  /api/v1/cards/{id}/activate:
    put:
      tags:
        - card-controller
      operationId: activateCard
      summary:
        "Activate card (ADMIN)"
      description: "Change card status to ACTIVE"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
  /api/v1/cards/my/{id}/request-block:
    put:
      tags:
        - card-controller
      operationId: blockMyCardRequest
      summary:
        "Request card block (USER)"
      description: "Change card status to REQUESTED_BLOCK"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
...
  /api/v1/cards:
    get:
      tags:
        - card-controller
      operationId: getCards
      summary:
        "All cards list (ADMIN)"
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 100
            minimum: 1
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - card-controller
      operationId: createCard
      summary:
        "Create new card (ADMIN)"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardChangeDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
...
  /api/v1/cards/{id}:
    get:
      tags:
        - card-controller
      operationId: getCard
      summary:
        "Return card by id (ADMIN)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - card-controller
      operationId: deleteCard
      summary:
        "Delete card by id (ADMIN)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]

    patch:
      tags:
        - card-controller
      operationId: createCard_1
      summary:
        "Update card by id (ADMIN)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardChangeDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
...
  /api/v1/users/{userId}/cards:
    get:
      tags:
        - card-controller
      operationId: getUserCards
      summary:
        "Return all user cards by user_id (ADMIN)"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 100
            minimum: 1
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string

      security:
        - bearerAuth: [ ]
...
  /api/v1/cards/my:
    get:
      tags:
        - card-controller
      operationId: getMyCards
      summary:
        "Return all current user cards (USER)"
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 100
            minimum: 1
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string

      security:
        - bearerAuth: [ ]
  /api/v1/cards/my/{id}:
    get:
      tags:
        - card-controller
      operationId: getMyCard
      summary:
        "Return current user card by card_id (USER)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string

      security:
        - bearerAuth: [ ]
  /api/v1/cards/my/{id}/balance:
    get:
      tags:
        - card-controller
      operationId: getMyCardBalance
      summary:
        "Return balance of current user card by card_id (USER)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: number
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
        '403':
          description: Forbidden
          content:
            '*/*':
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        balance:
          type: number
        expirationDate:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
            - REQUESTED_BLOCK
        ownerId:
          type: integer
          format: int64
...
    CardChangeDto:
      type: object
      properties:
        balance:
          type: number
        expirationDate:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
            - REQUESTED_BLOCK
        userId:
          type: integer
          format: int64
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer